+++ src/main.rs
@@ -3,8 +3,8 @@ mod ast;
 mod build_graph;
 mod compile;
 mod context;
-mod depfile;
 mod error;
+mod file;
 mod hash_type;
 mod infrastructure;
 mod ir;
